requirements.txt
Django==1.7.1
WebOb==1.4
bugsnag==2.0.2
dj-database-url==0.3.0
dj-static==0.0.6
django-braces==1.4.0
django-toolbelt==0.0.1
gunicorn==19.1.1
httplib2==0.9
numpy==1.9.1
pandas==0.15.1
path.py==7.0
psycopg2==2.5.4
python-dateutil==2.3
pytz==2014.9
requests==2.4.3
scikit-learn==0.14.1
scipy==0.14.0
six==1.8.0
static3==0.5.1
statsmodels==0.6.1